refactor: 포맷팅 및 타입 변경, 테스트 코드 중복 제거
- 수입률 천 단위 쉼표 포맷팅
- 총 상금 타입 int -> long 변경
- 테스트 코드 중복 리팩토링

+++ src/main/java/lotto/constants/LottoConstants.java
@@ -9,5 +9,5 @@ public final class LottoConstants {
     public static final int FIVE = 5;
     public static final int TWO = 2;
     public static final int ONE = 1;
-    public static final int ZERO = 0;
+    public static final long ZERO = 0L;
 }
